fix(vapor): keep the hydration index off the public IR struct
Adding `parent_id`/`index` to `NextRefIRNode` makes cargo-semver-checks fail
with `constructible_struct_adds_field`: the struct is externally
constructible and the workspace version is unchanged, so the new fields are a
breaking change. Derive both in the generator instead.

`GenerateContext` is crate-private, so it records where every `ChildRef`/
`NextRef` landed (parent plus absolute rendered index) and `generate_next_ref`
reads the predecessor's entry to compute its own index. Emitted code is
identical to the IR-field version, so the pinned sibling-navigation and
snapshot expectations are unchanged.

The expression fallback scanner moves out of `context.rs` into a submodule so
the file, already above the 350-line limit, does not grow.
